moved UiDrawing to seperate UiDrawer class
