+++ app/src/main/java/com/group2/catan_android/fragments/PlayerScoresFragment.java
@@ -1,5 +1,6 @@
 package com.group2.catan_android.fragments;
 
+import android.media.Image;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -7,6 +8,7 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.group2.catan_android.R;
@@ -20,6 +22,11 @@ public class PlayerScoresFragment extends Fragment {
     private TextView thirdPlayer;
     private TextView fourthPlayer;
     private TextView[] playerScoreViews;
+    private ImageView[] activePlayerViews;
+    private ImageView firstPlayerActive;
+    private ImageView secondPlayerActive;
+    private ImageView thirdPlayerActive;
+    private ImageView fourthPlayerActive;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -41,14 +48,34 @@ public class PlayerScoresFragment extends Fragment {
         playerScoreViews[2] = thirdPlayer;
         playerScoreViews[3] = fourthPlayer;
 
+        setupActivePlayerIndicators(view);
+
         return view;
     }
 
-    public void updateScores(List<Player> players){
-        for(int i = 0; i < players.size(); i++){
-            String playerScore = players.get(i).getDisplayName() + ": " + String.valueOf(players.get(i).getVictoryPoints());
+    public void setupActivePlayerIndicators(View v){
+        firstPlayerActive = v.findViewById(R.id.firstPlayerActive);
+        secondPlayerActive = v.findViewById(R.id.secondPlayerActive);
+        thirdPlayerActive = v.findViewById(R.id.thirdPlayerActive);
+        fourthPlayerActive = v.findViewById(R.id.fourthPlayerActive);
+
+        activePlayerViews = new ImageView[]{firstPlayerActive,secondPlayerActive,thirdPlayerActive,fourthPlayerActive};
+
+        for(ImageView activePlayerView : activePlayerViews){
+            activePlayerView.setVisibility(View.INVISIBLE);
+        }
+    }
+
+    public void updateScores(List<Player> playerList, Player activePlayer){
+        for(int i = 0; i < playerList.size(); i++){
+            String playerScore = playerList.get(i).getDisplayName() + ": " + String.valueOf(playerList.get(i).getVictoryPoints());
             playerScoreViews[i].setText(playerScore);
-            playerScoreViews[i].setTextColor(players.get(i).getColor());
+            playerScoreViews[i].setTextColor(playerList.get(i).getColor());
+
+            activePlayerViews[i].setVisibility(View.INVISIBLE);
+            if(playerList.get(i).getInGameID() == activePlayer.getInGameID()){
+                activePlayerViews[i].setBackgroundColor(View.VISIBLE);
+            }
         }
     }
 }
