+++ app/src/main/res/layout/fragment_player_scores.xml
@@ -1,62 +1,105 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".fragments.PlayerScoresFragment">
 
     <TextView
-    android:id="@+id/firstPlayerScore"
-    android:textColor="@color/black"
-    android:gravity="end|center_vertical"
-    android:autoSizeTextType="uniform"
-    android:layout_width="0dp"
-    android:layout_height="0dp"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toTopOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHeight_percent="0.25"
-    />
+        android:id="@+id/firstPlayerScore"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_marginEnd="20dp"
+        android:autoSizeTextType="uniform"
+        android:gravity="end|center_vertical"
+        android:textColor="@color/black"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.25"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/secondPlayerScore"
-        android:textColor="@color/black"
-        android:gravity="end|center_vertical"
-        android:autoSizeTextType="uniform"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/firstPlayerScore"
+        android:layout_marginEnd="20dp"
+        android:autoSizeTextType="uniform"
+        android:gravity="end|center_vertical"
+        android:textColor="@color/black"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent="0.25"
-        />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/firstPlayerScore" />
 
     <TextView
         android:id="@+id/thirdPlayerScore"
-        android:textColor="@color/black"
-        android:gravity="end|center_vertical"
-        android:autoSizeTextType="uniform"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/secondPlayerScore"
+        android:layout_marginEnd="20dp"
+        android:autoSizeTextType="uniform"
+        android:gravity="end|center_vertical"
+        android:textColor="@color/black"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent="0.25"
-        />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/secondPlayerScore" />
 
     <TextView
         android:id="@+id/fourthPlayerScore"
-        android:textColor="@color/black"
-        android:gravity="end|center_vertical"
-        android:autoSizeTextType="uniform"
         android:layout_width="0dp"
         android:layout_height="0dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/thirdPlayerScore"
-        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginEnd="20dp"
+        android:autoSizeTextType="uniform"
+        android:gravity="end|center_vertical"
+        android:textColor="@color/black"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_percent="0.25"
-        />
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/thirdPlayerScore" />
+
+    <ImageView
+        android:id="@+id/firstPlayerActive"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:src="@drawable/active_player_indicator"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.25"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toEndOf="@+id/firstPlayerScore"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/secondPlayerActive"
+        android:layout_width="00dp"
+        android:layout_height="0dp"
+        android:src="@drawable/active_player_indicator"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.25"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toEndOf="@+id/secondPlayerScore"
+        app:layout_constraintTop_toBottomOf="@+id/firstPlayerActive" />
+
+    <ImageView
+        android:id="@+id/thirdPlayerActive"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:src="@drawable/active_player_indicator"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.25"
+        app:layout_constraintStart_toEndOf="@+id/thirdPlayerScore"
+        app:layout_constraintTop_toBottomOf="@+id/secondPlayerActive" />
+
+    <ImageView
+        android:id="@+id/fourthPlayerActive"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:src="@drawable/active_player_indicator"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHeight_percent="0.25"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toEndOf="@+id/fourthPlayerScore"
+        app:layout_constraintTop_toBottomOf="@+id/thirdPlayerActive" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
